+++ CHANGELOG.md
@@ -1,7 +1,16 @@
-## 0.0.1
+## 1.1.0
 
-* New Release!
+* Changed the parameter from `pageStartFrom` to `initialPage`
+* Add a new example
+* Loading initial widget and callback parameter
+* Error handling
+* Add all parameter in listview
+* Add index to `itemBuilder`
 
 ## 1.0.1
 
 * Change package name
+
+## 0.0.1
+
+* New Release!